:art: Restructuring of dataset code.
